Put all the new goodies from reloaded into the js lib
This lets enterprise display all automations and other
goodies.

For both HelmReleaseDetail and KustomizationDetail, I've taken the
body of the page, and pasted it into a component, so that it can be
embedded as a component instead of as a page. It's not very nicely
done.
